Issue: WSW-17509 MISRA Fixes for mlan_glue.c patch 2
[Issue]
Rule 10.1 Operands shall not be of an inappropriate essential type.
Rule 10.3 The value of an expression shall not be assigned to an object with a narrower essential type or of a different essential type category.
Rule 10.4 Both operands of an operator in which the usual arithmetic conversions are performed shall have the same essential type category.

[Fix]
Comply as per rules

[Test]
Build success on IAR ARMGCC Keil MCUX.
Scan and Connect as STA.
uAP start and external station connect and ping.
STA iperf TCP UDP Client Server

[Reviewers]
Sandip Dalvi, Supriya Inamdar, Abhinav Kulkarni, Sanjay Yadav

+++ incl/wifidriver/wifi-decl.h
@@ -270,7 +270,7 @@ struct wifi_scan_result
     bool phtinfo_ie_present; /*!< PHT INFO IE present info */
 
     bool wmm_ie_present; /*!< WMM IE present info */
-    uint8_t band;        /*!< Band info */
+    uint16_t band;       /*!< Band info */
 
     bool wps_IE_exist;                         /*!< WPS IE exist info */
     uint16_t wps_session;                      /*!< WPS session */
